readded the frontend

+++ src/vm/vm.cc
@@ -3,10 +3,11 @@
 #include <utility>
 #include "vm.hh"
 #include <sstream>
+#include <cstdint>
 
-long long column_to_ord(const std::string column) {
-    long long ord = 0;
-    for (long long i = 0; i < column.size(); i++) {
+uint64_t column_to_ord(const std::string column) {
+    uint64_t ord = 0;
+    for (uint64_t i = 0; i < column.size(); i++) {
         ord *= 26;
         ord += std::tolower(column[i]) - 'a';
     }
@@ -14,7 +15,7 @@ long long column_to_ord(const std::string column) {
     return ord;
 }
 
-Number::Number(const long long start_column, const long long start_line, const double value) {
+Number::Number(const uint64_t start_column, const uint64_t start_line, const double value) {
     this->data_type = DataType::NUMBER;
     this->value = value;
 
@@ -22,7 +23,7 @@ Number::Number(const long long start_column, const long long start_line, const d
     this->start_line = start_line;
 }
 
-String::String(const long long start_column, const long long start_line, const std::string value) {
+String::String(const uint64_t start_column, const uint64_t start_line, const std::string value) {
     this->data_type = DataType::STRING;
     this->value = value;
 
@@ -34,10 +35,10 @@ Scope::Scope() {
     this->assignments.clear();
 }
 
-void Scope::assign_cell(const std::string column, const long long row, const RuntimeValue value) {
+void Scope::assign_cell(const std::string column, const uint64_t row, const RuntimeValue value) {
     std::stringstream ss;
     ss << "c"; // c stands for cell assignment
-    for (long long i = 0; i < column.size(); i++) {
+    for (uint64_t i = 0; i < column.size(); i++) {
         ss << std::tolower(column[i]);
     }
 
@@ -45,14 +46,14 @@ void Scope::assign_cell(const std::string column, const long long row, const Run
     this->assignments[ss.str()] = value;
 }
 
-void Scope::assign_range(const std::string column1, const long long row1, const std::string column2, const long long row2, const RuntimeValue value) {
+void Scope::assign_range(const std::string column1, const uint64_t row1, const std::string column2, const uint64_t row2, const RuntimeValue value) {
     std::stringstream ss;
     ss << "r"; // r stands for range assignment
-    for (long long i = 0; i < column1.size(); i++) {
+    for (uint64_t i = 0; i < column1.size(); i++) {
         ss << std::tolower(column1[i]);
     }
 
-    for (long long i = 0; i < column2.size(); i++) {
+    for (uint64_t i = 0; i < column2.size(); i++) {
         ss << std::tolower(column2[i]);
     }
 
@@ -60,13 +61,13 @@ void Scope::assign_range(const std::string column1, const long long row1, const 
     this->assignments[ss.str()] = value;
 }
 
-RuntimeValue Scope::retrieve(const long long start_column, const long long start_row, const std::string column, const long long row) const {
+RuntimeValue Scope::retrieve(const uint64_t start_column, const uint64_t start_row, const std::string column, const uint64_t row) const {
     for (auto it = this->assignments.begin(); it != this->assignments.end(); it++) {
         if (it->first[0] == 'c') {
             // Cell assignment
             std::string column_scope = "";
-            long long row_scope = 0;
-            for (long long i = 1; i < it->first.size(); i++) {
+            uint64_t row_scope = 0;
+            for (uint64_t i = 1; i < it->first.size(); i++) {
                 if ('0' <= it->first[i] && it->first[i] <= '9') {
                     row_scope *= 10;
                     row_scope += it->first[i] - '0';
@@ -80,7 +81,7 @@ RuntimeValue Scope::retrieve(const long long start_column, const long long start
             }
 
             bool flag = true;
-            for (long long i = 0; i < column.size(); i++) {
+            for (uint64_t i = 0; i < column.size(); i++) {
                 if (std::tolower(column[i]) != column_scope[i]) {
                     flag = false;
                 }
@@ -94,13 +95,13 @@ RuntimeValue Scope::retrieve(const long long start_column, const long long start
         }
 
         // Range assignment
-        long long column1 = 0;
-        long long row1 = 0;
+        uint64_t column1 = 0;
+        uint64_t row1 = 0;
         
-        long long column2 = 0;
-        long long row2 = 0;
+        uint64_t column2 = 0;
+        uint64_t row2 = 0;
 
-        for (long long i = 1; i < it->first.size(); i++) {
+        for (uint64_t i = 1; i < it->first.size(); i++) {
             if ('0' <= it->first[i] && it->first[i] <= '9') {
                 if (column2 == 0) {
                     row1 *= 10;
@@ -118,8 +119,8 @@ RuntimeValue Scope::retrieve(const long long start_column, const long long start
             }
         }
 
-        long long column_idx = column_to_ord(column);
-        long long row_idx = row;
+        uint64_t column_idx = column_to_ord(column);
+        uint64_t row_idx = row;
 
         if (column1 <= column_idx && column_idx <= column2 && row1 <= row_idx && row_idx <= row2) {
             return it->second;
